+++ packages/fretboard/index.js
@@ -1,7 +1,12 @@
 /**
+ * This module has functions to create fretboards and query for instrument
+ * tunnings.
+ *
+ * __Currently is NOT part of the tonal distribution__
+ *
  * @module fretboard
  */
-import { get, keys } from 'tonal-dictionary'
+import { dictionary } from 'tonal-dictionary'
 import { fromSemitones } from 'tonal-interval'
 import { transpose } from 'tonal-transpose'
 import { pc } from 'tonal-note'
@@ -9,7 +14,8 @@ import { asArr, compact, map } from 'tonal-array'
 import { numeric } from 'tonal-range'
 import { includes } from 'tonal-pitchset'
 
-var DICT = require('./tunings.json')
+var DATA = require('./tunings.json')
+var dict = dictionary(DATA, function (s) { return s.split(' ') })
 
 /**
  * Given a tuning name, returns the notes of the strings in the open position
@@ -21,7 +27,7 @@ var DICT = require('./tunings.json')
  * fret.tuning('guitar') // => [ 'E2', 'A2', 'D3', 'G3', 'B3', 'E4' ]
  * fret.tuning('charango') // => [ 'G4', 'G4', 'C5', 'C5', 'E5', 'E4', 'A4', 'A4', 'E5', 'E5' ]
  */
-export var tuning = get(null, DICT)
+export var tuning = dict.get
 
 /**
  * Given a tuning name returns the notes of the strings in open position
@@ -49,7 +55,7 @@ export function simpleTuning (src) {
  * @param {Boolean} aliases - get aliases or not
  * @return {Array} an array of tuning names
  */
-export var names = keys(DICT)
+export var names = dict.keys
 
 /**
  * Build a fretboard using a given tuning (or tuning name), first and last